requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
djangorestframework==3.9.2
idna==2.8
Pillow==5.3.0
pytz==2018.9
sqlparse==0.3.0
djangorestframework-jwt==1.11.0
